requirements.txt
Django==5.2.9
gunicorn
numpy
pandas
scikit-learn
joblib
plotly
rdkit-pypi
playwright
kaleido
reportlab
